Assignment Second completed

+++ Controller/userController.js
@@ -1,7 +1,9 @@
 
 var bcrypt = require('bcrypt');
 var params = require('params');
-var users = require('../Model/usermodel.js')
+var users = require('../Model/usermodel.js');
+var jwt = require('jsonwebtoken');
+var secret = 'ThisIsMySecretKey';
 
 function validator(req,res,next){
     if(req.body.Username === ''){
@@ -18,7 +20,7 @@ function validator(req,res,next){
 
 function genHash(req,res,next){
 
-    var saltRounds = 10;
+    var saltRounds = 10;    
     bcrypt.hash(req.body.Password, saltRounds, function(err, hash){
     
         if(hash){
@@ -77,31 +79,74 @@ function checkUserAlreadyExists(req,res,next){
     })
 }
 
+
+
+
+// var token = req.headers.authorization;
+// token = token.slice(7, token.length).trimLeft();
+// jwt.verify(token,secret,
+//     function(err,result){
+//         console.log(err,result);
+
+//         if(result){
+//             console.log("Correct token");
+//             next;
+//         }
+//         if(err){
+//             res.json({err});
+//         }
+//     })
+//     next;
+
 function deleteUser(req,res,next){
     if(req.params.id === null || req.params.id === undefined){
         res.status(404);
         res.json({status:404, message:'Id is not provided'})
     }
-    users.users.destroy({
-        where:{
-            id:req.params.id
-        }
-    })
-    .then(function(result){
-        console.log(result);
-        if(result === 0){
-            res.status(500);
-            res.json({status:409, message:"couldnot delete user."})
-        }
-        res.status(200);
-        res.json({status:201, message:"User delete successfully."})
-    })
+
+    if(req.headers.authorization === undefined){
+        res.status(401);
+        res.json({status:404, message:"Unauthorized Access"});
+    }
+
+    var token = req.headers.authorization;
+    token = token.slice(7, token.length).trimLeft();
+    jwt.verify(token,secret,
+        function(err,result){
+            console.log(err,result);
+
+            if(result){
+                users.users.destroy({
+                    where:{
+                        id:req.params.id
+                    }
+                })
+                .then(function(result){
+                    console.log(result);
+                    if(result === 0){
+                        res.status(500);
+                        res.json({status:409, message:"Couldnot delete user."})
+                    }
+                    res.status(200);
+                    res.json({status:201, message:"User delete successfully."})
+                })
+            
+            
+                .catch(function(err){
+                    console.log(err)
+                    res.json(err);
+                })
+                next;
+            }
+            if(err){
+                res.json({err});
+            }
+        })
+        next;
 
 
-    .catch(function(err){
-        console.log(err)
-        res.json(err);
-    })
+
+    
 }
